+++ lib/getISO3166.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const iso3166_data = require('./iso3166_data');
+const iso3166_data = require('./iso3166Data');
 
 // assign default to USA country first
