+++ lib/getISO3166ByPhone.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const iso3166_data = require('./iso3166_data');
+const iso3166_data = require('./iso3166Data');
 
 module.exports = function getISO3166ByPhone(phone) {
 	return iso3166_data
